pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sdf-sampler"
version = "0.1.0"
description = "An example project using various Python libraries"
authors = [
  { name = "Michael Kofler", email = "michael.kofler@tuwien.ac.at" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
  "numpy",
  "matplotlib",
  "gustaf",
  "scikit-image",
  "libigl",
  "trimesh",
  "embreex"
]

[project.optional-dependencies]
develop = [
  "black",
  "pytest"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["SdfSampler", "SdfSampler.*"]